+++ src/tests/fixtures/mpd_client.rs
@@ -20,6 +20,7 @@ use crate::mpd::{
         Update,
         Volume,
         list::MpdList,
+        list_all::ListAll,
         list_playlist::FileList,
         metadata_tag::MetadataTagExt,
         mpd_config::MpdConfig,
@@ -179,6 +180,11 @@ impl MpdClient for TestMpdClient {
             ValueChange::Decrease(val) => self.volume.dec_by(val),
             ValueChange::Set(val) => self.volume.set_value(val),
         };
+        Ok(())
+    }
+
+    fn crossfade(&mut self, seconds: u32) -> MpdResult<()> {
+        self.status.xfade = Some(seconds);
         Ok(())
     }
 
@@ -468,7 +474,7 @@ impl MpdClient for TestMpdClient {
         todo!("Not yet implemented")
     }
 
-    fn list_all(&mut self, _: Option<&str>) -> MpdResult<LsInfo> {
+    fn list_all(&mut self, _: Option<&str>) -> MpdResult<ListAll> {
         todo!("Not yet implemented")
     }