+++ src/ui/mod.rs
@@ -470,6 +470,22 @@ impl<'ui> Ui<'ui> {
                         Ok(())
                     });
                 }
+                GlobalAction::CrossfadeUp => {
+                    let current_xfade = ctx.status.xfade.unwrap_or(0);
+                    let new_xfade = current_xfade.saturating_add(1);
+                    ctx.command(move |client| {
+                        client.crossfade(new_xfade)?;
+                        Ok(())
+                    });
+                }
+                GlobalAction::CrossfadeDown => {
+                    let current_xfade = ctx.status.xfade.unwrap_or(0);
+                    let new_xfade = current_xfade.saturating_sub(1);
+                    ctx.command(move |client| {
+                        client.crossfade(new_xfade)?;
+                        Ok(())
+                    });
+                }
                 GlobalAction::SeekForward
                     if matches!(ctx.status.state, State::Play | State::Pause) =>
                 {
@@ -1002,19 +1018,6 @@ impl Config {
 
     fn as_text_style(&self) -> ratatui::style::Style {
         self.theme.text_color.map(|color| Style::default().fg(color)).unwrap_or_default()
-    }
-
-    fn as_styled_progress_bar(&self) -> widgets::progress_bar::ProgressBar<'_> {
-        let progress_bar_colors = &self.theme.progress_bar;
-        widgets::progress_bar::ProgressBar::default()
-            .elapsed_style(progress_bar_colors.elapsed_style)
-            .thumb_style(progress_bar_colors.thumb_style)
-            .track_style(progress_bar_colors.track_style)
-            .start_char(&self.theme.progress_bar.symbols[0])
-            .elapsed_char(&self.theme.progress_bar.symbols[1])
-            .thumb_char(&self.theme.progress_bar.symbols[2])
-            .track_char(&self.theme.progress_bar.symbols[3])
-            .end_char(&self.theme.progress_bar.symbols[4])
     }
 
     fn as_styled_scrollbar(&self) -> Option<ratatui::widgets::Scrollbar<'_>> {